+++ api/src/main/java/jakarta/data/exceptions/EmptyResultException.java
@@ -15,34 +15,38 @@
  *
  *  SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data;
+package jakarta.data.exceptions;
 
-public class NonUniqueResultException extends DataException {
+/**
+ * Data access exception thrown when a result was expected to have at least one row (or element)
+ * but zero rows (or elements) were actually returned.
+ */
+public class EmptyResultException extends DataException {
 
     /**
-     * Constructs a new NonUniqueResultException exception with the specified detail message.
+     * Constructs a new EmptyResultException exception with the specified detail message.
      *
      * @param message the detail message.
      */
-    public NonUniqueResultException(String message) {
+    public EmptyResultException(String message) {
         super(message);
     }
 
     /**
-     * Constructs a new NonUniqueResultException exception with the specified detail message.
+     * Constructs a new EmptyResultException exception with the specified detail message.
      *
      * @param message the detail message.
      */
-    public NonUniqueResultException(String message, Throwable cause) {
+    public EmptyResultException(String message, Throwable cause) {
         super(message, cause);
     }
 
     /**
-     * Constructs a new NonUniqueResultException exception with the specified cause.
+     * Constructs a new EmptyResultException exception with the specified cause.
      *
      * @param cause the cause.
      */
-    public NonUniqueResultException(Throwable cause) {
+    public EmptyResultException(Throwable cause) {
         super(cause);
     }
 }
