+++ api/src/main/java/module-info.java
@@ -18,5 +18,6 @@
 module jakarta.data.api {
     exports jakarta.data;
     exports jakarta.data.repository;
+    exports jakarta.data.exceptions;
     opens jakarta.data.repository;
 }
